Update to only notify complete orders since last notification
