+++ skills/csharp/experimental/FoodOrderSkill/FoodOrderSkill/Dialogs/RepeatOrderDialog.cs
@@ -2,9 +2,9 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.InteropServices.ComTypes;
 using System.Threading;
 using System.Threading.Tasks;
 using AdaptiveCards;
@@ -12,12 +12,9 @@ using Alexa.NET.Response;
 using Bot.Builder.Community.Adapters.Alexa.Attachments;
 using FoodOrderSkill.MockBackEnd;
 using FoodOrderSkill.Models;
-using Microsoft.Azure.CognitiveServices.ContentModerator.Models;
-using Microsoft.Azure.CognitiveServices.Language.LUIS.Runtime.Models;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Dialogs;
 using Microsoft.Bot.Builder.Dialogs.Choices;
-using Microsoft.Bot.Builder.Solutions.Responses;
 using Microsoft.Bot.Schema;
 using Newtonsoft.Json.Linq;
 
@@ -28,9 +25,13 @@ namespace FoodOrderSkill.Dialogs
         // TODO: Remove and replace with service to hit favoriteMeals API
         private BackEndDB localDB;
 
+        // temp storage for pro-active scenario
+        private ConcurrentDictionary<string, ConversationReference> _conversationReferences;
+
         public RepeatOrderDialog(
             IServiceProvider serviceProvider,
-            IBotTelemetryClient telemetryClient)
+            IBotTelemetryClient telemetryClient,
+            ConcurrentDictionary<string, ConversationReference> conversationReferences)
             : base(nameof(RepeatOrderDialog), serviceProvider, telemetryClient)
         {
             var steps = new WaterfallStep[]
@@ -43,6 +44,7 @@ namespace FoodOrderSkill.Dialogs
             };
 
             this.localDB = new BackEndDB();
+            _conversationReferences = conversationReferences;
 
             AddDialog(new WaterfallDialog(nameof(RepeatOrderDialog), steps));
             AddDialog(new ChoicePrompt(DialogIds.FavoriteOrderPrompt));
@@ -271,12 +273,21 @@ namespace FoodOrderSkill.Dialogs
                 orderPlacedActivity.Attachments.Add(alexaCardAttachment);
 
                 await stepContext.Context.SendActivityAsync(orderPlacedActivity);
+
+                AddConversationReference(stepContext.Context.Activity);
+
                 return await stepContext.EndDialogAsync();
             }
             else
             {
                 return await stepContext.ReplaceDialogAsync(nameof(RepeatOrderDialog), null, cancellationToken);
             }
+        }
+
+        private void AddConversationReference(Activity activity)
+        {
+            var conversationReference = activity.GetConversationReference();
+            _conversationReferences.AddOrUpdate($"{conversationReference.User.Id}_{activity.ChannelId}", conversationReference, (key, newValue) => conversationReference);
         }
 
         private class DialogIds
